Create ShellScript_To_Copy_File_Btw_Folders.sh
This script can be used to copy file from one location(Secondary location) to Primary location.

Case:  ODS-CB gets debit card file from source system(iCMS) to a file server(To a specific folder called ICMS). But this system has been migrated to new System called webCMS , so we created a folder called WCMS in the file server and job expects the file to be present in WCMS instead of ICMS. However, source system had problem to transfer file to latest folder and we were asked to point our jobs to old folder. 
We can't Change our jobs as it requires lot of effort and complete end to end testing. So, I have developed this script that moves file from ICMS to WCMS, in doing so, it checks all scenarios and tells about the status of the file.

Script is self explanatory.